requirements.txt
asgiref==3.8.1
autopep8==2.3.1
boto==2.49.0
boto3==1.35.54
botocore==1.35.54
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
crispy-bootstrap4==2024.10
dj-database-url==2.3.0
Django==5.1.2
django-crispy-forms==2.3
django-storages==1.14.4
docutils==0.21.2
gunicorn==23.0.0
idna==3.10
jmespath==1.0.1
packaging==24.1
pillow==11.0.0
psycopg2==2.9.10
pycodestyle==2.12.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
s3transfer==0.10.3
six==1.16.0
sqlparse==0.5.1
stripe==11.2.0
typing_extensions==4.12.2
urllib3==2.2.3
